Display, Add & Delete Students enhancement inside Beauty School Web App - Part 2

+++ src/main/java/com/marlena/beautyschool/repository/PersonRepository.java
@@ -4,9 +4,11 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import com.marlena.beautyschool.model.Person;
 
+import javax.persistence.Id;
+
 
 @Repository
-public interface PersonRepository extends CrudRepository<Person, Long> {
+public interface PersonRepository extends CrudRepository<Person, Integer> {
 
     Person readByEmail(String email);
 }
